rename Oscillators to Oscillator as it now produces single oscillators
