+++ js/awe/Prometheus/Oscillator/Controller.js
