+++ js/awe/Prometheus/Oscillator/Model.js
